Move ValidatedTextInput to react-components
